net: Expose make_server to other crates via test-util feature
This patch extracts make_server() (and related helpers) from net/tests/main.rs
into a reusable module, guarded by a new test-util feature.

With this change, other Servo crates, especially libservo, can spin up a
lightweight http://localhost server during integration tests without pulling
the full net subsystem into normal builds.

This is needed for upcoming WebView integration tests that require real
origins (e.g., http://localhost) so that storage APIs can be exercised
properly when testing upcoming SiteDataManager
