+++ components/net/lib.rs
@@ -20,6 +20,8 @@ pub mod protocols;
 pub mod request_interceptor;
 pub mod resource_thread;
 pub mod subresource_integrity;
+#[cfg(feature = "test-util")]
+pub mod test_util;
 mod websocket_loader;
 
 /// An implementation of the [Fetch specification](https://fetch.spec.whatwg.org/)
