+++ components/servo/tests/webview.rs
@@ -10,6 +10,10 @@ use std::rc::Rc;
 
 use dpi::PhysicalSize;
 use euclid::Point2D;
+use http_body_util::combinators::BoxBody;
+use hyper::body::{Bytes, Incoming};
+use hyper::{Request as HyperRequest, Response as HyperResponse};
+use net::test_util::{make_body, make_server};
 use servo::{
     ContextMenuAction, ContextMenuElementInformation, ContextMenuElementInformationFlags,
     ContextMenuItem, Cursor, EmbedderControl, InputEvent, InputMethodType, JSValue,
@@ -91,6 +95,38 @@ fn test_create_webview() {
     let url = webview.url();
     assert!(url.is_some());
     assert_eq!(url.unwrap().to_string(), "about:blank");
+}
+
+#[test]
+fn test_create_webview_http() {
+    let servo_test = ServoTest::new();
+
+    static MESSAGE: &'static [u8] = b"<!DOCTYPE html>\nHello";
+    let handler =
+        move |_: HyperRequest<Incoming>,
+              response: &mut HyperResponse<BoxBody<Bytes, hyper::Error>>| {
+            *response.body_mut() = make_body(MESSAGE.to_vec());
+        };
+    let (server, url) = make_server(handler);
+
+    let delegate = Rc::new(WebViewDelegateImpl::default());
+
+    let webview = WebViewBuilder::new(servo_test.servo(), servo_test.rendering_context.clone())
+        .delegate(delegate.clone())
+        .url(url.into_url())
+        .build();
+
+    servo_test.spin(move || !delegate.url_changed.get());
+
+    let _ = server.close();
+
+    let url = webview.url();
+    assert!(url.is_some());
+    let url = url.unwrap();
+    assert_eq!(url.scheme(), "http");
+    let host = url.host_str();
+    assert!(host.is_some());
+    assert_eq!(host.unwrap(), "localhost");
 }
 
 #[test]
